refactor: move the login and user session related code into the login folder
The code in the oauth.ts file was highly related to the code in the
login folder. It made no sense to have it in the setup folder.

Furthermore, the configuration/index.ts file also contained highly login
specific code like the creation and configuration of the user manager.

This commit also simplifies the mocked authentication. There is no real
need to implement a stripped down version of the user manager, because
the only thing the silentSignin method did was a call to the onSessionRenewed
hook with static data.
